Refactoring classe astratta FormActivity Rinominata classe astratta FormActivity in ActivityWithDialogs, estratto metodo checkInput() e inserito in un'interfaccia denominata FormActivity.
